Board state and logic finished
I decide to store board state as a EnumSet<Cell>, because it will be easier to store state and display board.

I represent one element of board as a Cell (which is an Enum), thanks to that future operations on
the board state, will not need to pass the board references.

Cell has a empty element. With empty being 0 indices are from 1, which suits user input better.

Sign also has a empty element. It is put to differentiate Cells taken by X or O from Cell without sign.